feat: refactored decryption of user data for sessions
