+++ utils/API/index.js
@@ -1,4 +1,5 @@
 export * from "./amount";
+export * from "./decryptUserData";
 export * from "./response";
 export * from "./skip";
 export * from "./take";
